+++ app/sprinkles/core/locale/it_IT/messages.php
@@ -16,47 +16,47 @@
 return [
     "@PLURAL_RULE" => 1,
 
-    "ABOUT" => "Riguardo a noi",
+    "ABOUT" => "Su di noi",
 
 	"CAPTCHA" => [
 	    "@TRANSLATION" => "Captcha",
-        "FAIL" => "Domanda di sicurezza sbagliata",
-        "SPECIFY" => "Inserire il captcha",
-        "VERIFY" => "Verifica la captcha"
+        "FAIL" => "Captcha errato",
+        "SPECIFY" => "Inserisci il captcha",
+        "VERIFY" => "Verifica captcha"
     ],
 
-    "CSRF_MISSING" => "Sigillo CSRF mancante. Prova a aggiornare la pagina e poi di inviarlo nuovamente?",
+    "CSRF_MISSING" => "Token CSRF mancante. Prova ad aggiornare la pagina e poi invia nuovamente la richiesta.",
 
-    "DB_INVALID" => "Impossibile connettersi al database. Se sei un amministratore, controlla il registro PHP o UserFrosting.",
+    "DB_INVALID" => "Impossibile connettersi al database. Se sei un amministratore, controlla il log PHP o di UserFrosting.",
     "DESCRIPTION"   => "Descrizione",
     "DOWNLOAD" => [
-        "@TRANSLATION" => "Scaricare",
+        "@TRANSLATION" => "Scarica",
         "CSV" => "Scarica CSV"
     ],
 
     "EMAIL" => [
-        "@TRANSLATION" => "E-mail",
+        "@TRANSLATION" => "Email",
         "YOUR" => "La tua email"
     ],
 
-    "HOME"  => "Inizio",
+    "HOME"  => "Home",
 
     "LEGAL" => [
-        "@TRANSLATION" => "Politica Legale",
+        "@TRANSLATION" => "Politica legale",
         "DESCRIPTION" => "La nostra politica legale si applica al tuo utilizzo di questo sito e dei nostri servizi."
     ],
 
     "LOCALE" => [
-        "@TRANSLATION" => "Località"
+        "@TRANSLATION" => "Lingua"
     ],
 
     "NAME"  => "Nome",
     "NAVIGATION" => "Navigazione",
-    "NO_RESULTS" => "Spiacenti, non abbiamo niente qui.",
+    "NO_RESULTS" => "Spiacenti, qui non c'è niente.",
 
     "PAGINATION" => [
         "GOTO" => "Vai alla pagina",
-        "SHOW" => "Mostrare",
+        "SHOW" => "Mostra",
 
         // Paginator
         // possible variables: {size}, {page}, {totalPages}, {filteredPages}, {startRow}, {endRow}, {filteredRows} and {totalRows}
@@ -69,55 +69,55 @@ return [
     ],
     "PRIVACY" => [
         "@TRANSLATION" => "Politica sulla riservatezza",
-        "DESCRIPTION" => "La nostra politica sulla privacy descrive quali tipi di informazioni raccoglieremo da te e come lo useremo."
+        "DESCRIPTION" => "La nostra politica sulla privacy descrive quali tipi di informazioni raccoglieremo da te e come le useremo."
     ],
 
     "SLUG" => "Slug",
     "SLUG_CONDITION" => "Slug/Condizioni",
     "SLUG_IN_USE" => "Esiste già uno slug <strong>{{slug}}</strong>",
     "STATUS" => "Stato",
-    "SUGGEST" => "Suggerire",
+    "SUGGEST" => "Suggerisci",
 
     "UNKNOWN" => "Sconosciuto",
 
     // Actions words
     "ACTIONS"  => "Azioni",
-    "ACTIVATE" => "Attivare",
+    "ACTIVATE" => "Attiva",
     "ACTIVE"   => "Attivo",
-    "ADD"      => "Aggiungere",
+    "ADD"      => "Aggiungi",
     "CANCEL"   => "Annulla",
     "CONFIRM"  => "Conferma",
-    "CREATE"   => "Creare",
+    "CREATE"   => "Crea",
     "DELETE"   => "Elimina",
-    "DELETE_CONFIRM" => "Sei sicuro di voler cancellare questo?",
-    "DELETE_CONFIRM_YES" => "Sì, elimini",
+    "DELETE_CONFIRM" => "Sei sicuro di volerlo eliminare?",
+    "DELETE_CONFIRM_YES" => "Sì, elimina",
     "DELETE_CONFIRM_NAMED" => "Sei sicuro di voler eliminare {{name}}?",
-    "DELETE_CONFIRM_YES_NAMED" => "Sì, eliminare {{name}}",
+    "DELETE_CONFIRM_YES_NAMED" => "Sì, elimina {{name}}",
     "DELETE_CANNOT_UNDONE" => "Questa azione non può essere annullata.",
     "DELETE_NAMED" => "Elimina {{name}}",
-    "DENY" => "Nega",   
-    "DISABLE"   => "Disattivare",
+    "DENY" => "Rifiuta",
+    "DISABLE"   => "Disabilita",
     "DISABLED"  => "Disabilitato",
     "EDIT"      => "Modifica",
-    "ENABLE"    => "Abilitare",
+    "ENABLE"    => "Abilita",
     "ENABLED"   => "Abilitato",
-    "OVERRIDE"  => "Alterare",
-    "RESET"     => "Azzerare",
-    "SAVE"      => "Memorizzare",
-    "SEARCH"    => "Cercare",
-    "SORT"      => "Ordinare",
-    "SUBMIT"    => "Inviare",
+    "OVERRIDE"  => "Sovrascrivi",
+    "RESET"     => "Azzera",
+    "SAVE"      => "Salva",
+    "SEARCH"    => "Cerca",
+    "SORT"      => "Ordina",
+    "SUBMIT"    => "Invia",
 	"SUCCESS"   => "Successo",
-    "PRINT"     => "Stampare",
-    "REMOVE"    => "Rimuovere",
-    "UNACTIVATED" => "Non attivato",
-    "UPDATE"    => "Aggiornare",
+    "PRINT"     => "Stampa",
+    "REMOVE"    => "Rimuovi",
+    "UNACTIVATED" => "Disattivato",
+    "UPDATE"    => "Aggiorna",
     "YES"       => "Sì",
     "NO"        => "No",
     "OPTIONAL"  => "Opzionale",
 
     // Misc.
-    "BUILT_WITH_UF" => "Construito <a href=\"http://www.userfrosting.com\">UserFrosting</a>",
-    "ADMINLTE_THEME_BY" => "Tema da <strong><a href=\"http://almsaeedstudio.com\">Almsaeed Studio</a>.</strong> Tutti i diritti riservati",
-    "WELCOME_TO" => "Benvenuto a {{title}}!"
+    "BUILT_WITH_UF" => "Construito con <a href=\"http://www.userfrosting.com\">UserFrosting</a>",
+    "ADMINLTE_THEME_BY" => "Tema di <strong><a href=\"http://almsaeedstudio.com\">Almsaeed Studio</a>.</strong> Tutti i diritti riservati",
+    "WELCOME_TO" => "Benvenuto su {{title}}!"
 ];