+++ CHANGELOG.md
@@ -5,8 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## v4.1.19
+- Prevent setup to run again if already configured when using `bake`
+- Fix `Unique::getPaginatedQuery` to call to `addSelect` instead of `select` during the pre-paginated query
+- Updated Spanish Translation (Thanks @silvioq !)
+- Fix error template in WhoopsRenderer (#885; Thanks @silvioq !)
+
+## v4.1.18
+- Bakery setup wizard for SMTP config + separate SMTP setup in it's own command (https://github.com/userfrosting/UserFrosting/issues/874)
+- Update Italian translations (https://github.com/userfrosting/UserFrosting/pull/875)
+- Return deleted `row` in `ufCollection` event (https://github.com/userfrosting/UserFrosting/pull/873)
+
 ## v4.1.17-alpha
-- Lock `gulp-uf-bundle-assets` at v2.28.0 until Silic0nS0ldier/gulp-uf-bundle-assets#5 is resolved (see #859)
+- Lock `gulp-uf-bundle-assets` at v2.28.0 until userfrosting/gulp-uf-bundle-assets#5 is resolved (see #859)
 - Add missing getInfo methods for GroupController and RoleController (#837)
 
 ## v4.1.16-alpha